+++ torchbearer/callbacks/cutout.py
@@ -1,7 +1,9 @@
 import torchbearer
 from torchbearer import Callback
+
 import torch
-import numpy as np
+from torch.distributions import Beta
+
 from torchbearer.bases import cite
 
 cutout = """
@@ -19,6 +21,16 @@ random_erase = """
   author={Zhong, Zhun and Zheng, Liang and Kang, Guoliang and Li, Shaozi and Yang, Yi},
   journal={arXiv preprint arXiv:1708.04896},
   year={2017}
+}
+"""
+
+
+cutmix = """
+@article{yun2019cutmix,
+  title={Cutmix: Regularization strategy to train strong classifiers with localizable features},
+  author={Yun, Sangdoo and Han, Dongyoon and Oh, Seong Joon and Chun, Sanghyuk and Choe, Junsuk and Yoo, Youngjoon},
+  journal={arXiv preprint arXiv:1905.04899},
+  year={2019}
 }
 """
 
@@ -48,11 +60,16 @@ class Cutout(Callback):
     """
     def __init__(self, n_holes, length, constant=0., seed=None):
         super(Cutout, self).__init__()
-        self.cutter = BatchCutout(n_holes, length, constant=constant, seed=seed)
+        self.constant = constant
+        torch.manual_seed(seed)
+        self.cutter = BatchCutout(n_holes, length, length)
 
     def on_sample(self, state):
         super(Cutout, self).on_sample(state)
-        state[torchbearer.X] = self.cutter(state[torchbearer.X])
+        mask = self.cutter(state[torchbearer.X])
+        erase_locations = mask == 0
+        constant = torch.ones_like(state[torchbearer.X]) * self.constant
+        state[torchbearer.X][erase_locations] = constant[erase_locations]
 
 
 @cite(random_erase)
@@ -80,11 +97,72 @@ class RandomErase(Callback):
     """
     def __init__(self, n_holes, length, seed=None):
         super(RandomErase, self).__init__()
-        self.cutter = BatchCutout(n_holes, length, seed=seed, random_erase=True)
+        torch.manual_seed(seed)
+        self.cutter = BatchCutout(n_holes, length, length)
 
     def on_sample(self, state):
         super(RandomErase, self).on_sample(state)
-        state[torchbearer.X] = self.cutter(state[torchbearer.X])
+        mask = self.cutter(state[torchbearer.X])
+        erase_locations = mask == 0
+        random = torch.rand_like(state[torchbearer.X])
+        state[torchbearer.X][erase_locations] = random[erase_locations]
+
+
+@cite(cutmix)
+class CutMix(Callback):
+    """ Cutmix callback which replaces a random patch of image data with the corresponding patch from another image.
+    This callback also converts labels to one hot before combining them according to the lambda parameters, sampled from
+    a beta distribution as is done in the paper.
+
+    Example: ::
+
+        >>> from torchbearer import Trial
+        >>> from torchbearer.callbacks import CutMix
+
+        # Example Trial which does Cutout regularisation
+        >>> erase = CutMix(1, classes=10)
+        >>> trial = Trial(None, callbacks=[erase], metrics=['acc'])
+
+    Args:
+        alpha (float): The alpha value for the beta distribution.
+        classes (int): The number of classes for conversion to one hot.
+        seed: Random seed
+
+    State Requirements:
+        - :attr:`torchbearer.state.X`: State should have the current data stored
+        - :attr:`torchbearer.state.Y_TRUE`: State should have the current data stored
+    """
+    def __init__(self, alpha, classes=-1, seed=None):
+        super(CutMix, self).__init__()
+        self.classes = classes
+        torch.manual_seed(seed)
+        self.dist = Beta(torch.tensor([alpha]), torch.tensor([alpha]))
+
+    def _to_one_hot(self, target):
+        if target.dim() == 1:
+            target = target.unsqueeze(1)
+            one_hot = torch.zeros_like(target).repeat(1, self.classes)
+            one_hot.scatter_(1, target, 1)
+            return one_hot
+        return target
+
+    def on_sample(self, state):
+        super(CutMix, self).on_sample(state)
+        lam = self.dist.sample()
+        length = (1 - lam).sqrt()
+        cutter = BatchCutout(1, (length * state[torchbearer.X].size(-1)).round().item(), (length * state[torchbearer.X].size(-2)).round().item())
+        mask = cutter(state[torchbearer.X])
+        erase_locations = mask == 0
+        permutation = torch.randperm(state[torchbearer.X].size(0))
+
+        state[torchbearer.X][erase_locations] = state[torchbearer.X][permutation][erase_locations]
+
+        target = self._to_one_hot(state[torchbearer.TARGET]).float()
+        state[torchbearer.TARGET] = lam * target + (1 - lam) * target[permutation]
+
+    def on_sample_validation(self, state):
+        super(CutMix, self).on_sample_validation(state)
+        state[torchbearer.TARGET] = self._to_one_hot(state[torchbearer.TARGET]).float()
 
 
 class BatchCutout(object):
@@ -92,15 +170,14 @@ class BatchCutout(object):
 
     Args:
         n_holes (int): Number of patches to cut out of each image.
-        length (int): The length (in pixels) of each square patch.
+        width (int): The width (in pixels) of each square patch.
+        height (int): The height (in pixels) of each square patch.
         seed: Random seed
     """
-    def __init__(self, n_holes, length, constant=0., random_erase=False, seed=None):
+    def __init__(self, n_holes, width, height):
         self.n_holes = n_holes
-        self.length = length
-        self.random_erasing = random_erase
-        self.constant = constant
-        np.random.seed(seed)
+        self.width = width
+        self.height = height
 
     def __call__(self, img):
         """
@@ -115,29 +192,20 @@ class BatchCutout(object):
         h = img.size(-2)
         w = img.size(-1)
 
-        mask = np.ones((b, h, w), np.float32)
+        mask = torch.ones((b, h, w), device=img.device)
 
         for n in range(self.n_holes):
-            y = np.random.randint(h, size=b)
-            x = np.random.randint(w, size=b)
+            y = torch.randint(h, (b,)).long()
+            x = torch.randint(w, (b,)).long()
 
-            y1 = np.clip(y - self.length // 2, 0, h)
-            y2 = np.clip(y + self.length // 2, 0, h)
-            x1 = np.clip(x - self.length // 2, 0, w)
-            x2 = np.clip(x + self.length // 2, 0, w)
+            y1 = (y - self.height // 2).clamp(0, h)
+            y2 = (y + self.height // 2).clamp(0, h)
+            x1 = (x - self.width // 2).clamp(0, w)
+            x2 = (x + self.width // 2).clamp(0, w)
 
             for batch in range(b):
                 mask[batch, y1[batch]: y2[batch], x1[batch]: x2[batch]] = 0
 
-        mask = torch.from_numpy(mask).unsqueeze(1).repeat(1, c, 1, 1)
+        mask = mask.unsqueeze(1).repeat(1, c, 1, 1)
 
-        erase_locations = mask == 0
-
-        if self.random_erasing:
-            random = torch.from_numpy(np.random.rand(*img.shape)).to(torch.float)
-        else:
-            random = torch.from_numpy(np.ones_like(img)).to(torch.float) * self.constant
-
-        img[erase_locations] = random[erase_locations]
-
-        return img
+        return mask